Automatically add the dependencies label to Renovate PRs (#35)
There are some inconsistencies in how the Renovate config is applied to
different repos. Some reference the primary configuration set forth in
this repo, and others have their own configuration. In either case, most
of the Renovate configs don't contain the directive to add the
'dependencies' label to the PRs created by the Renovate bot, and because
of this, they are candidates for the [stalebot
pruning](https://github.com/powerhome/software/blob/main/modules/github-repo/stale.yml.tpl)
and will be closed after 4 days of inactivity.

There are a few instances of this happening in
[microsites](https://github.com/powerhome/microsites/pull/188#issuecomment-2576446669),
[pac](https://github.com/powerhome/pac/pull/2388#issuecomment-2588087865),
[connect-apple](https://github.com/powerhome/connect-apple/pull/1170#issuecomment-2588124874),
and others.