Add config to ignore reviewdog version updates (#40)
Similar to https://github.com/powerhome/renovate-config/pull/36, we want
to ignore individual repos trying to update the software-repo-controlled
reviewdog actions.

https://github.com/powerhome/example-rails-app/pull/1129 was merged this
morning, and will cause the drift check to fail because it was updated
outside of a Terraform apply.

https://github.com/powerhome/software/pull/1459 is the PR that should
_actually_ have been merged to get this updated, but there are likely
other repos that could be affected, and this warrants a global config to
ignore these Actions updates.